chain fee should be a float

+++ api/typeDefs/chainFee.js
@@ -2,6 +2,6 @@ import { gql } from 'graphql-tag'
 
 export default gql`
   extend type Query {
-    chainFee: Int!
+    chainFee: Float!
   }
 `
